Move nav to its own component
